Explicitly specify the highest header processed
Since some protocols can have some valid fields with null values, there should be a workaround (This was particularly prevalent with Rhizome parsing since the type field is either 1 or 0).
Instead, the parsing function will explicitly return the highest-parsed header for ease of interpretation by the rest of the program.